# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Salud360 – API de Integración de Citas Médicas
  description: |
    API REST para el registro de citas médicas en la red Salud360.

    ## Patrones de integración implementados
    - **Point-to-Point Channel**: La solicitud de facturación se envía a `billing.queue`
      y es procesada por un único consumidor (sistema de facturación).
    - **Publish/Subscribe Channel**: El evento `CITA_CONFIRMADA` se publica en el exchange
      `appointments.events` (fanout) y es recibido por `notifications.queue` y `analytics.queue`.
    - **Transferencia de archivos**: Se genera un registro en `data/outbox/auditoria-citas.csv`
      para el sistema legado de auditoría.

    ## Manejo de errores
    Los errores de validación y de integración se registran en `data/errors/citas-rechazadas.log`.
  version: 1.0.0
  contact:
    name: Equipo de Integración Salud360
    email: integracion@salud360.ec

servers:
  - url: http://localhost:8080
    description: Servidor local de desarrollo

tags:
  - name: Citas Médicas
    description: Endpoints para el registro y consulta de citas médicas

paths:

  /api/citas:

    post:
      tags: [Citas Médicas]
      summary: Registrar una nueva cita médica
      description: |
        Recibe los datos de una solicitud de cita, los valida y, si son correctos,
        inicia el flujo de integración completo:
        1. Envía mensaje de facturación a `billing.queue` (Point-to-Point).
        2. Publica evento `CITA_CONFIRMADA` en `appointments.events` (Publish/Subscribe).
        3. Genera línea en `auditoria-citas.csv` para el sistema legado.
      operationId: registrarCita
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CitaRequest'
            examples:
              citaValida:
                summary: "Solicitud válida completa"
                value:
                  idCita: "CITA-1001"
                  paciente: "Ana Torres"
                  correo: "ana.torres@email.com"
                  especialidad: "Cardiología"
                  fechaCita: "2026-06-15"
                  sede: "Centro Norte"
                  valor: 45.50
              citaInvalida:
                summary: "Solicitud inválida (campos faltantes)"
                value:
                  idCita: ""
                  paciente: ""
                  correo: "no-es-un-email"
                  especialidad: ""
                  fechaCita: ""
                  sede: ""
                  valor: -10
      responses:
        '202':
          description: Cita aceptada — flujo de integración iniciado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitaAceptada'
              example:
                status: "ACEPTADA"
                idCita: "CITA-1001"
                mensaje: "Cita registrada y flujo de integración iniciado correctamente"
        '400':
          description: Datos de la solicitud inválidos o incompletos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorValidacion'
              example:
                status: "RECHAZADA"
                idCita: "N/A"
                errores:
                  - "paciente: El campo paciente es obligatorio"
                  - "correo: El campo correo es obligatorio"
        '422':
          description: Error de validación de negocio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorNegocio'
              example:
                status: "RECHAZADA"
                idCita: "CITA-1001"
                error: "La fechaCita [2020-01-01] no puede ser una fecha pasada"
        '500':
          description: Error interno al procesar la integración

    get:
      tags: [Citas Médicas]
      summary: Listar citas aceptadas
      description: Devuelve las citas que han sido aceptadas y procesadas durante la sesión actual.
      operationId: listarCitas
      responses:
        '200':
          description: Lista de citas aceptadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListaCitas'

components:
  schemas:

    CitaRequest:
      type: object
      required: [idCita, paciente, correo, especialidad, fechaCita, sede, valor]
      properties:
        idCita:
          type: string
          description: Identificador único de la cita
          example: "CITA-1001"
        paciente:
          type: string
          description: Nombre completo del paciente
          example: "Ana Torres"
        correo:
          type: string
          format: email
          description: Correo electrónico del paciente
          example: "ana.torres@email.com"
        especialidad:
          type: string
          description: Especialidad médica
          example: "Cardiología"
        fechaCita:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
          description: Fecha de la cita (formato YYYY-MM-DD)
          example: "2026-06-15"
        sede:
          type: string
          description: Sede del centro médico
          example: "Centro Norte"
        valor:
          type: number
          format: double
          minimum: 0.01
          description: Valor de la cita en USD (mayor a 0)
          example: 45.50

    CitaAceptada:
      type: object
      properties:
        status:
          type: string
          example: "ACEPTADA"
        idCita:
          type: string
          example: "CITA-1001"
        mensaje:
          type: string
          example: "Cita registrada y flujo de integración iniciado correctamente"

    ErrorValidacion:
      type: object
      properties:
        status:
          type: string
          example: "RECHAZADA"
        idCita:
          type: string
          example: "N/A"
        errores:
          type: array
          items:
            type: string
          example:
            - "paciente: El campo paciente es obligatorio"
            - "correo: El campo correo es obligatorio"

    ErrorNegocio:
      type: object
      properties:
        status:
          type: string
          example: "RECHAZADA"
        idCita:
          type: string
          example: "CITA-1001"
        error:
          type: string
          example: "La fechaCita no puede ser una fecha pasada"

    ListaCitas:
      type: object
      properties:
        total:
          type: integer
          example: 1
        citas:
          type: array
          items:
            $ref: '#/components/schemas/CitaRequest'